LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/scroll-to-fragid/fragment-and-encoding-expected.txt


FAIL U+00FF should find U+00FF assert_greater_than: #%C3%BF expected a number greater than 1000 but got 0
FAIL Percent-encoded UTF-8 BOM should find U+FEFF as BOM is not stripped when decoding assert_greater_than: #%EF%BB%BF expected a number greater than 1000 but got 0
FAIL %FF should not find U+00FF as decoding it gives U+FFFD assert_equals: #%FF expected 0 but got 9416
PASS Valid UTF-8 + invalid UTF-8 should not be matched to the utf8-decoded former + the isomorphic-decoded latter 
